+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 * Fix toggling debugging to always send the `debug`/`nodebug` messages after enabling/disabling the breakpoints
 * Fix call hierarchy "Show Incoming Calls" and "Show Outgoing Calls" context menu items
 * Fix "Go to Symbol in Editor..." and "Go to Symbol in Workspace..." support to parse multi-line predicate scope directives
+* Fix "Go to Symbol in Editor..." and "Go to Symbol in Workspace..." support when entity, predicate, and non-terminal names contain single quotes
 * Fix "go to Symbol in Editor..." parsing of predicates and non-terminals with the same name but different arity
 * Fix code navigation commands to reject calls from inside line comments and support cancellation